feat: Add an endpoint to get messages
